Express Templating Demo
These are all the files that are being used or required.
Here we have learned about templating and used a little bit of reddit fake data to grasp over the concept of templating.
Also we have learned about partials where we do not need to write the same code in every page to display the same things.